Refactoring of the tasks and gulpfile.babel.js !
