--- tasks/errorHandler.js
@@ -1,17 +1,0 @@
-/*
- * @title Error Handler
- */
-
-// Depedencies
-import notifier from 'node-notifier';
-
-// Task
-export function errorHandler(error) {
-  notifier.notify({
-    title: 'Gulp Error',
-    message: error.message,
-    timeout: 3
-  });
-  console.error('\x1b[31m', error.message ,'\x1b[0m');
-  this.emit('end');
-}
